converted all contact api endpoints to private and tested the apis

+++ backend/controllers/contactController.js
@@ -5,16 +5,16 @@ const Contact = require("../models/contactModel");
 
 //desc  ->  get all contacts
 //route ->  GET /api/contacts
-//access ->  public
+//access ->  private
 
 const getContacts = asyncHandler( async (req,res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id: req.user.id});
     res.status(200).json(contacts);
 });
 
 //desc  ->  create new contact
 //route ->  POST /api/contacts
-//access ->  public
+//access ->  private
 
 const createContact = asyncHandler( async (req,res) => {
     console.log("The request body is ",req.body);
@@ -25,6 +25,7 @@ const createContact = asyncHandler( async (req,res) => {
     }
 
     const contact = await Contact.create({
+        user_id: req.user.id,
         name,
         email,
         phone
@@ -35,7 +36,7 @@ const createContact = asyncHandler( async (req,res) => {
 
 //desc  ->  get a contact
 //route ->  GET /api/contacts/:id
-//access ->  public
+//access ->  private
 
 const getContact = asyncHandler( async (req,res) => {
     const contact = await Contact.findById(req.params.id);
@@ -48,13 +49,18 @@ const getContact = asyncHandler( async (req,res) => {
 
 //desc  ->  update a contact
 //route ->  PUT /api/contacts/:id
-//access ->  public
+//access ->  private
 
 const updateContact = asyncHandler( async (req,res) => {
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact Not Found!");
+    }
+
+    if(contact.user_id.toString() !== req.user.id){
+       res.status(403);
+       throw new Error("User don't have permission to update other user contacts");
     }
 
     const updatedContact = await Contact.findByIdAndUpdate(
@@ -67,7 +73,7 @@ const updateContact = asyncHandler( async (req,res) => {
 
 //desc  ->  delete a contact
 //route ->  DELETE /api/contacts/:id
-//access ->  public
+//access ->  private
 
 const deleteContact = asyncHandler( async (req,res) => {
     const contact = await Contact.findById(req.params.id);
@@ -76,6 +82,11 @@ const deleteContact = asyncHandler( async (req,res) => {
         throw new Error("Contact Not Found!");
     }
 
+    if(contact.user_id.toString() !== req.user.id){
+        res.status(403);
+        throw new Error("User don't have permission to delete other user contacts");
+    }
+    
     await Contact.findByIdAndDelete(req.params.id);
     res.status(201).json(contact);
 });
